Move rest of GUT/ and GUTI/ to universe/
