+++ src/universe/GUTAnnotatedTypeFactory.java
@@ -1,4 +1,4 @@
-package GUT;
+package universe;
 
 import java.lang.annotation.Annotation;
 import java.util.Arrays;
