+++ src/universe/GUTChecker.java
@@ -1,4 +1,4 @@
-package GUT;
+package universe;
 
 
 import javax.annotation.processing.SupportedOptions;
