+++ src/universe/GUTIChecker.java
@@ -1,4 +1,4 @@
-package GUTI;
+package universe;
 
 import javax.annotation.processing.SupportedOptions;
