+++ src/universe/GUTIConstraintSerializer.java
@@ -1,4 +1,4 @@
-package GUTI;
+package universe;
 
 import java.lang.reflect.Constructor;
