+++ src/universe/GUTIMaxSatSerializer.java
@@ -1,4 +1,4 @@
-package GUTI;
+package universe;
 
 import java.util.ArrayList;
 import java.util.List;
