+++ src/universe/GUTIVisitor.java
@@ -1,4 +1,4 @@
-package GUTI;
+package universe;
 
 import java.util.List;
 import java.util.Set;
