+++ src/universe/GUTViewpointAdaptor.java
@@ -1,4 +1,4 @@
-package GUT;
+package universe;
 
 
 import javax.lang.model.element.AnnotationMirror;
